+++ source/standalone_examples/api/omni.isaac.ros_bridge/rtx_lidar.py
@@ -17,6 +17,7 @@ from omni.isaac.core.utils.extensions import enable_extension
 from omni.isaac.core import SimulationContext
 from omni.isaac.core.utils import viewports, stage, nucleus
 from omni.syntheticdata import sensors
+import omni.kit.viewport.utility
 from pxr import Gf
 
 # enable ROS bridge extension
@@ -36,12 +37,10 @@ if not rosgraph.is_master_online():
 
 
 # Create a new viewport for the RTX sensor and acquire the viewport window
-vpi = omni.kit.viewport_legacy.get_viewport_interface()
-instance = vpi.create_instance()
-viewport_handle = vpi.get_instance("Viewport 2")
-viewport = vpi.get_viewport_window(viewport_handle)
+window = omni.kit.viewport.utility.create_viewport_window("Viewport 2")
 # in order for the sensor to generate data properly we let the viewport know that it should create a buffer for the associated render variable.
-viewport.add_aov("RtxSensorCpu", False)
+omni.kit.viewport.utility.add_aov_to_viewport(window.viewport_api, "RtxSensorCpu")
+
 
 # Locate Isaac Sim assets folder to load environment and robot stages
 assets_root_path = nucleus.get_assets_root_path()