+++ source/extensions/omni.isaac.internal_tools/docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+
+## [1.1.0] - 2022-09-01
+
+### Changed
+- Remove legacy viewport calls
+
+
 ## [1.0.0] - 2022-06-13
 
 ### Added