+++ source/extensions/omni.isaac.isaac_sensor/docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.1.1] - 2022-09-01
+
+### Changed
+- Remove legacy viewport calls from tests
+
 ## [1.1.0] - 2022-08-24
 
 ### Added