+++ source/extensions/omni.isaac.ros2_bridge/docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+
+## [1.6.0] - 2022-08-30
+
+### Changed
+- Remove direct legacy viewport calls
 ## [1.5.0] - 2022-08-15
 
 ### Added